Polished unit tests
Made unit_tests output look more pleasing by using Pandas DataFrames for table printing. In addition: 1. Made some messages from the unit tests easier to understand 2. Modified the behavior of insert to prevent insertion of duplicate items 3. Modified behavior of the test for insert in order to make it work with the new version of insert() 4. Placed the code to check for duplicates in its own function _check_for_duplicates